+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+-   Report error if test continues after returning
+-   Report skipped tests correctly
+
 ## [3.5.3] 2020-04-21
 
 -   Escape attribute values